Added css to the header and linked the css to the html.
In style.css, set box-sizing property inside the
universal selector to border-box. This declared
that the box-model properties of the html elements
would be included in their height and width. Also,
added css reset for margin and padding inside the
universal selector to clear all the default
padding and margin for the html elements. Added
font "Roboto" for the body with "Helvetica Neue",
Arial, sans-serif as fallbacks. Then styled the
list items, by removing default list style from
which removed bullet points in front of the links
and removed default text decoration from links to
remove the underline from underneath them.