Add logo and nav links to the header of index.html
Add blank image and empty links as unordered list items
inside the header tags in index.html. For the links,
set target to "_blank" so that the links open in the
new tab, and rel attribute to "noopener noreferrer"
where noopener prevents the new page from using the
window.opener property to manipulate the original
page and noreferrer prevents browser from sending
referrer information so the destination site does not
know where the visitor came from.