Add a font-family to the body and style the links.
Add 'Roboto' font to elements inside of the body along with
sans-serif and serif font-family as fall-backs. Add a background
color to the header. Remove default list styling for navigation
links by changing the default "disc" to "none". Change the nav
links font size to 18px from default 16px, and color to white
from default blue. And add yellow color to the links when hover
and on keyboard navigation using tab.