Add the box model properties and the flexbox properties to the call to action section.
Add block margin, padding, and border radius to the call to action
section. Set the call to action section as a flex container with
a gap of 80px between its flex items (text and button) and
use justify-content to align the contents to have space
between them in terms of the call to action section and use
align-self to align it centrally across the inline axis in
terms of the body.